requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
openai>=1.61.0
sqlalchemy==2.0.34
masumi==0.1.30
degen_crew @ git+https://github.com/idopterlabs/masumi-cardano-agent
